+++ languages/de/default.php
@@ -4,5 +4,10 @@
 /**
  * Content elements
  */
-$GLOBALS['TL_LANG']['CTE']['newselement']    = array('Nachricht', 'Zeigt eine News an.');
+$GLOBALS['TL_LANG']['CTE']['newselement']    = ['Nachricht', 'Zeigt eine News an.'];
+
+/**
+ * Frontend Modules
+ */
+$GLOBALS['TL_LANG']['FMD']['newselement']    = ['Nachricht', 'Zeigt eine News an.'];